tests/wpt/web-platform-tests/html/canvas/element/the-canvas-state/2d.state.saverestore.stackdepth.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.state.saverestore.stackdepth</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.state.saverestore.stackdepth</h1>
<p class="desc">save()/restore() stack depth is not unreasonably limited</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() stack depth is not unreasonably limited");
_addTest(function(canvas, ctx) {

var limit = 512;
for (var i = 1; i < limit; ++i)
{
    ctx.save();
    ctx.lineWidth = i;
}
for (var i = limit-1; i > 0; --i)
{
    _assertSame(ctx.lineWidth, i, "ctx.lineWidth", "i");
    ctx.restore();
}


});
</script>
