LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/the-canvas-state/2d.state.saverestore.stackdepth-expected.txt

2d.state.saverestore.stackdepth
save()/restore() stack depth is not unreasonably limited
Actual output:

PASS save()/restore() stack depth is not unreasonably limited
